# Backend/docs/swagger.yaml
definitions:
...
  errors.APIError:
    properties:
      error:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
  models.AdminUser:
    properties:
      created_at:
        type: string
      department:
        $ref: '#/definitions/models.Department'
      department_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      last_login:
        type: string
      name:
        type: string
      phone:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      status:
        type: integer
      updated_at:
        type: string
    type: object
  models.Department:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Department'
        type: array
      created_at:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Menu:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Menu'
        type: array
      component_path:
        type: string
      created_at:
        type: string
      icon:
        type: string
      id:
        type: integer
      parent_id:
        type: integer
      path:
        type: string
      permission_items:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      permissions:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      sort:
        type: integer
      status:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.Permission:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      menu:
        $ref: '#/definitions/models.Menu'
      menu_id:
        type: integer
      role_ids:
        items:
          type: integer
        type: array
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      menus:
        items:
          $ref: '#/definitions/models.Menu'
        type: array
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.AdminUser'
        type: array
    type: object
host: localhost:8082
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: 这是GOCOM项目的API文档，提供了详细的接口说明和使用方法。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GOCOM API
  version: "1.0"
paths:
...
  /api/menus:
    get:
      consumes:
      - application/json
      description: 获取所有菜单项，支持分页
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回菜单列表、总数、页码和每页数量
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取所有菜单
      tags:
      - 菜单管理
    post:
      consumes:
      - application/json
      description: 创建一个新的菜单项
      parameters:
      - description: 菜单信息
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "201":
          description: 返回创建的菜单信息
          schema:
            $ref: '#/definitions/models.Menu'
        "400":
          description: 无效的请求参数
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建新菜单
      tags:
      - 菜单管理
    put:
      consumes:
      - application/json
      description: 更新特定的菜单项
      parameters:
      - description: 更新的菜单信息
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: 返回更新后的菜单信息
          schema:
            $ref: '#/definitions/models.Menu'
        "400":
          description: 无效的请求参数
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新菜单
      tags:
      - 菜单管理
  /api/menus/{id}:
    delete:
      consumes:
      - application/json
      description: 删除特定的菜单项
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功消息
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 无效的ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除菜单
      tags:
      - 菜单管理
    get:
      consumes:
      - application/json
      description: 通过ID获取特定的菜单项
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回菜单信息
          schema:
            $ref: '#/definitions/models.Menu'
        "400":
          description: 无效的ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 菜单未找到
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 通过ID获取菜单
      tags:
      - 菜单管理
  /api/menus/{id}/permissions:
    get:
      consumes:
      - application/json
      description: 获取指定菜单ID下的所有权限数据
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回权限列表
          schema:
            items:
              $ref: '#/definitions/models.Permission'
            type: array
        "400":
          description: 无效的菜单ID
          schema:
            $ref: '#/definitions/errors.APIError'
        "404":
          description: 菜单未找到
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.APIError'
      summary: 获取指定菜单的所有权限
      tags:
      - 菜单管理
  /api/menus/getRoutes:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的菜单列表
      produces:
      - application/json
      responses:
        "200":
          description: 返回用户的菜单列表
          schema:
            items:
              $ref: '#/definitions/models.Menu'
            type: array
        "401":
          description: 用户未认证
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.APIError'
      summary: 获取用户菜单
      tags:
      - 菜单管理
  /api/menus/menusWithRoles:
    get:
      consumes:
      - application/json
      description: 获取所有菜单项及其关联的角色，支持分页
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回菜单列表（包含角色）、总数、页码和每页数量
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.APIError'
      summary: 获取所有菜单及其角色
      tags:
      - 菜单管理
  /api/menus/permissions/{id}:
    post:
      consumes:
      - application/json
      description: 为指定菜单添加一组权限
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限列表
        in: body
        name: permissions
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Permission'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 添加权限成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/errors.APIError'
        "404":
          description: 菜单未找到
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.APIError'
      summary: 为菜单添加权限
      tags:
      - 菜单管理
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 请输入 'Bearer {token}' 来进行身份验证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"